requirements.txt
# F2 Portfolio Recommender Agent - Dependencies (Restructured)
# Core AI/LLM - Cerebras API
cerebras-cloud-sdk>=1.0.0
langchain>=0.1.9
langchain-community>=0.0.24

# Portfolio Optimization & Data Analysis
PyPortfolioOpt>=1.5.5
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Guardrails & Safety
guardrails-ai>=0.4.0
pydantic>=2.6.0

# Testing & Evaluation
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Frontend & Visualization
streamlit>=1.32.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.1
jupyter>=1.0.0
